Centralize intid cleanup in api.delete
