+++ src/bika/lims/api/__init__.py
@@ -77,7 +77,7 @@ from senaite.core.interfaces import ITemporaryObject
 from z3c.form.validator import Data as ValidatorData
 from zope import globalrequest
 from zope.annotation.interfaces import IAttributeAnnotatable
-from zope.component import createObject
+from zope.component import getAllUtilitiesRegisteredFor
 from zope.component import getUtility
 from zope.component import queryMultiAdapter
 from zope.container.contained import notifyContainerModified
@@ -88,7 +88,7 @@ from zope.interface import Invalid
 from zope.interface import alsoProvides
 from zope.interface import directlyProvides
 from zope.interface import noLongerProvides
-from zope.lifecycleevent import ObjectCreatedEvent
+from zope.intid.interfaces import IIntIds
 from zope.lifecycleevent import ObjectMovedEvent
 from zope.publisher.browser import TestRequest
 from zope.schema import getFieldsInOrder
@@ -526,6 +526,23 @@ def catalog_object(obj, recursive=False):
             catalog_object(child, recursive=recursive)
 
 
+def unregister_intid(obj):
+    """Unregister the object from all IIntIds utilities
+
+    Useful when deleting an object without firing IObjectRemovedEvent,
+    so its intid keyref does not survive as an orphan in storage.
+
+    :param obj: object to unregister
+    :type obj: ATContentType/DexterityContentType
+    """
+    for intids in getAllUtilitiesRegisteredFor(IIntIds):
+        try:
+            intids.unregister(obj)
+        except KeyError:
+            # not registered with this utility
+            pass
+
+
 def delete(obj, check_permissions=True, suppress_events=False):
     """Deletes the given object
 
@@ -540,8 +557,14 @@ def delete(obj, check_permissions=True, suppress_events=False):
 
     # un-catalog the object from all catalogs (uid_catalog included)
     uncatalog_object(obj)
+    if suppress_events:
+        # five.intid's removal subscriber listens on IObjectRemovedEvent,
+        # which is suppressed here — drop the keyref manually to avoid an
+        # orphan entry pinning the dead oid in storage.
+        unregister_intid(obj)
+
     # delete the object
-    parent = get_parent(obj)
+    parent = aq_parent(obj)
     parent._delObject(obj.getId(), suppress_events=suppress_events)
